+++ src/fuellib/__init__.py
@@ -1,0 +1,5 @@
+"""fuellib."""
+
+from . import comp, data, decomp, fuel, gcm
+
+__all__ = ["comp", "data", "decomp", "fuel", "gcm"]