+++ src/specfuel/__init__.py
@@ -2,6 +2,6 @@
 
 __version__ = "0.0.0"
 
-from . import data, fuel, gcm, inv
+from . import inv
 
-__all__ = ["data", "fuel", "gcm", "inv"]
+__all__ = ["inv"]